requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.8
Django==3.2.6
django-appconf==1.0.4
django-cryptography==1.0
django-encrypted-id-cryptography==1.1.0
django-environ==0.5.0
django-filter==2.4.0
django-mysql==4.0.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
mysql==0.0.3
Pillow==8.3.1
pycparser==2.20
pycryptodomex==3.10.1
pytz==2021.1
sqlparse==0.4.1
